Boards and shit

+++ BattleShipsCPP/AIPlayer.cpp
@@ -1,4 +1,4 @@
-#include "AIPlayer.h"
+#include "GamePlayer.h"
 #include <stdlib.h>
 
 using namespace std;
@@ -7,7 +7,7 @@ AIPlayer::AIPlayer()
 {
 	srand(time(NULL));
 
-	int namesCount = size(PossibleNames);
+	size_t namesCount = size(PossibleNames);
 	int nameIndex = rand() % namesCount;
 	_name = PossibleNames[nameIndex];
 }
